+++ src/main/java/com/dabomstew/pkrandom/constants/Gen5Constants.java
@@ -53,7 +53,8 @@ public class Gen5Constants {
     public static final int bw1StarterTextOffset = 18, bw1CherenText1Offset = 26, bw1CherenText2Offset = 53;
 
     public static final int bw1StriatonCityGymTextOffset = 2, bw1StriatonCityGymText2Offset = 8,
-        bw1NacreneCityGymTextOffset = 1;
+        bw1NacreneCityGymTextOffset = 1, bw1CasteliaCityBurghTextOffset = 0,
+        bw1CasteliaCityBurghText2Offset = 5, bw1CasteliaCityBurghText3Offset = 11;
     
     public static final int bw1CasteliaCityItemTextOffset = 3, bw1CasteliaCityItemText2Offset = 7,
         bw1CasteliaCityItemText3Offset = 11, bw1CasteliaCityPraiseTextOffset = 6;