+++ ChangeLog.md
@@ -55,4 +55,5 @@
    * Nacrene City Gym Guy will use true type and appropriate weakness when using type-theming on Gym Leaders, or use generic terms when completely random
    * Castelia City item guy will use correct starter names when giving type-boosting items. Items will not match the types of the new starters
    * Castelia City cafe guy will use correct types when referring to Striaton City leaders, or use generic terms when completely random
+   * Castelia City Gym Trainers and Burgh will use the true type of the gym when type-theming, or use generic terms when completely random.
  * Added `Normalize` and `Stall` to list of negative abilities that can be banned